[BPK-4383]: Migrate datatable and datepicker
